+++ languagemodels/config.py
@@ -167,6 +167,7 @@ models = [
     {
         "name": "all-MiniLM-L6-v2",
         "tuning": "embedding",
+        "query_prefix": "",
         "params": 22e6,
         "quantization": "int8",
         "backend": "ct2",
@@ -176,6 +177,17 @@ models = [
     {
         "name": "gte-small",
         "tuning": "embedding",
+        "query_prefix": "",
+        "params": 33e6,
+        "quantization": "int8",
+        "backend": "ct2",
+        "architecture": "encoder-only-transformer",
+        "license": "mit",
+    },
+    {
+        "name": "bge-small-en",
+        "tuning": "embedding",
+        "query_prefix": "Represent this sentence for searching relevant passages: ",
         "params": 33e6,
         "quantization": "int8",
         "backend": "ct2",
@@ -185,6 +197,7 @@ models = [
     {
         "name": "e5-small-v2",
         "tuning": "embedding",
+        "query_prefix": "",
         "params": 33e6,
         "quantization": "int8",
         "backend": "ct2",