+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Automatically use query prefixes as needed for embeddings
+- Add phi-1.5 model
 - Add support for gte-small embeddings
 - Add support for bge-small-en embeddings